고수찾기 별점 표시기능 적용 - 정인

+++ MeetGo/src/main/java/com/kh/meetgo/gosu/model/dto/GosuOpt.java
@@ -26,8 +26,8 @@ public class GosuOpt {
     private int estimateCount;
     private Date createDate;
     private int reviewCount;
-    private int totalRevPoint;
-    private int avgRevPoint;
+    private double totalRevPoint;
+    private double avgRevPoint;
 
 	// 나중에 별점 추가
 }
